♻️ refactor: separate the plugin service to plugin service and tool service
