+++ src/services/__tests__/tool.test.ts
@@ -1,13 +1,8 @@
-import { LobeChatPluginManifest } from '@lobehub/chat-plugin-sdk';
 import { Mock, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { PluginModel } from '@/database/client/models/plugin';
-import { DB_Plugin } from '@/database/client/schemas/plugin';
 import { globalHelpers } from '@/store/global/helpers';
-import { LobeTool } from '@/types/tool';
-import { LobeToolCustomPlugin } from '@/types/tool/plugin';
 
-import { InstallPluginParams, pluginService } from '../plugin';
+import { toolService } from '../tool';
 import openAPIV3 from './openai/OpenAPI_V3.json';
 import OpenAIPlugin from './openai/plugin.json';
 
@@ -18,21 +13,12 @@ vi.mock('@/store/global/helpers', () => ({
     getCurrentLanguage: vi.fn(),
   },
 }));
-vi.mock('@/database/client/models/plugin', () => ({
-  PluginModel: {
-    getList: vi.fn(),
-    create: vi.fn(),
-    delete: vi.fn(),
-    update: vi.fn(),
-    clear: vi.fn(),
-  },
-}));
 
 beforeEach(() => {
   vi.resetAllMocks();
 });
 
-describe('PluginService', () => {
+describe('ToolService', () => {
   describe('getPluginList', () => {
     it('should fetch and return the plugin list', async () => {
       // Arrange
@@ -45,7 +31,7 @@ describe('PluginService', () => {
       ) as any;
 
       // Act
-      const pluginList = await pluginService.getPluginList();
+      const pluginList = await toolService.getPluginList();
 
       // Assert
       expect(globalHelpers.getCurrentLanguage).toHaveBeenCalled();
@@ -60,7 +46,7 @@ describe('PluginService', () => {
       global.fetch = vi.fn(() => Promise.reject(new Error('Network error')));
 
       // Act & Assert
-      await expect(pluginService.getPluginList()).rejects.toThrow('Network error');
+      await expect(toolService.getPluginList()).rejects.toThrow('Network error');
     });
   });
 
@@ -112,7 +98,7 @@ describe('PluginService', () => {
         }),
       ) as any;
 
-      const manifest = await pluginService.getPluginManifest(manifestUrl);
+      const manifest = await toolService.getPluginManifest(manifestUrl);
 
       expect(fetch).toHaveBeenCalledWith(manifestUrl);
       expect(manifest).toEqual(fakeManifest);
@@ -120,7 +106,7 @@ describe('PluginService', () => {
 
     it('should return error on noManifest', async () => {
       try {
-        await pluginService.getPluginManifest();
+        await toolService.getPluginManifest();
       } catch (e) {
         expect(e).toEqual(new TypeError('noManifest'));
       }
@@ -138,7 +124,7 @@ describe('PluginService', () => {
       ) as any;
 
       try {
-        await pluginService.getPluginManifest(manifestUrl);
+        await toolService.getPluginManifest(manifestUrl);
       } catch (e) {
         expect(e).toEqual(new TypeError('manifestInvalid'));
       }
@@ -149,7 +135,7 @@ describe('PluginService', () => {
       global.fetch = vi.fn(() => Promise.reject(new Error('Network error')));
 
       try {
-        await pluginService.getPluginManifest(manifestUrl);
+        await toolService.getPluginManifest(manifestUrl);
       } catch (e) {
         expect(e).toEqual(new TypeError('fetchError'));
       }
@@ -170,7 +156,7 @@ describe('PluginService', () => {
       ) as any;
 
       try {
-        await pluginService.getPluginManifest(manifestUrl);
+        await toolService.getPluginManifest(manifestUrl);
       } catch (e) {
         expect(e).toEqual(new TypeError('urlError'));
       }
@@ -188,7 +174,7 @@ describe('PluginService', () => {
       ) as any;
 
       try {
-        await pluginService.getPluginManifest(manifestUrl);
+        await toolService.getPluginManifest(manifestUrl);
       } catch (e) {
         expect(e).toEqual(new TypeError('fetchError'));
       }
@@ -228,7 +214,7 @@ describe('PluginService', () => {
           }),
         ) as any;
 
-        const manifest = await pluginService.getPluginManifest(manifestUrl);
+        const manifest = await toolService.getPluginManifest(manifestUrl);
 
         expect(manifest).toMatchSnapshot();
       });
@@ -266,7 +252,7 @@ describe('PluginService', () => {
         ) as any;
 
         try {
-          await pluginService.getPluginManifest(manifestUrl);
+          await toolService.getPluginManifest(manifestUrl);
         } catch (e) {
           expect(e).toEqual(new TypeError('openAPIInvalid'));
         }
@@ -274,141 +260,8 @@ describe('PluginService', () => {
     });
   });
 
-  describe('installPlugin', () => {
-    it('should install a plugin', async () => {
-      // Arrange
-      const fakePlugin = {
-        identifier: 'test-plugin',
-        manifest: { name: 'TestPlugin', version: '1.0.0' } as unknown as LobeChatPluginManifest,
-        type: 'plugin',
-      } as InstallPluginParams;
-      vi.mocked(PluginModel.create).mockResolvedValue(fakePlugin);
-
-      // Act
-      const installedPlugin = await pluginService.installPlugin(fakePlugin);
-
-      // Assert
-      expect(PluginModel.create).toHaveBeenCalledWith(fakePlugin);
-      expect(installedPlugin).toEqual(fakePlugin);
-    });
-  });
-
-  describe('getInstalledPlugins', () => {
-    it('should return a list of installed plugins', async () => {
-      // Arrange
-      const fakePlugins = [{ identifier: 'test-plugin', type: 'plugin' }] as LobeTool[];
-      vi.mocked(PluginModel.getList).mockResolvedValue(fakePlugins as DB_Plugin[]);
-
-      // Act
-      const installedPlugins = await pluginService.getInstalledPlugins();
-
-      // Assert
-      expect(PluginModel.getList).toHaveBeenCalled();
-      expect(installedPlugins).toEqual(fakePlugins);
-    });
-  });
-
-  describe('uninstallPlugin', () => {
-    it('should uninstall a plugin', async () => {
-      // Arrange
-      const identifier = 'test-plugin';
-      vi.mocked(PluginModel.delete).mockResolvedValue();
-
-      // Act
-      const result = await pluginService.uninstallPlugin(identifier);
-
-      // Assert
-      expect(PluginModel.delete).toHaveBeenCalledWith(identifier);
-      expect(result).toBe(undefined);
-    });
-  });
-
-  describe('createCustomPlugin', () => {
-    it('should create a custom plugin', async () => {
-      // Arrange
-      const customPlugin = {
-        identifier: 'custom-plugin',
-        manifest: {},
-        type: 'customPlugin',
-      } as LobeToolCustomPlugin;
-      vi.mocked(PluginModel.create).mockResolvedValue(customPlugin);
-
-      // Act
-      const result = await pluginService.createCustomPlugin(customPlugin);
-
-      // Assert
-      expect(PluginModel.create).toHaveBeenCalledWith({
-        ...customPlugin,
-        type: 'customPlugin',
-      });
-      expect(result).toEqual(customPlugin);
-    });
-  });
-
-  describe('updatePlugin', () => {
-    it('should update a plugin', async () => {
-      // Arrange
-      const id = 'plugin-id';
-      const value = { settings: { ab: '1' } } as unknown as LobeToolCustomPlugin;
-      vi.mocked(PluginModel.update).mockResolvedValue(1);
-
-      // Act
-      const result = await pluginService.updatePlugin(id, value);
-
-      // Assert
-      expect(PluginModel.update).toHaveBeenCalledWith(id, value);
-      expect(result).toEqual(1);
-    });
-  });
-
-  describe('updatePluginManifest', () => {
-    it('should update a plugin manifest', async () => {
-      // Arrange
-      const id = 'plugin-id';
-      const manifest = { name: 'NewPluginManifest' } as unknown as LobeChatPluginManifest;
-      vi.mocked(PluginModel.update).mockResolvedValue(1);
-
-      // Act
-      const result = await pluginService.updatePluginManifest(id, manifest);
-
-      // Assert
-      expect(PluginModel.update).toHaveBeenCalledWith(id, { manifest });
-      expect(result).toEqual(1);
-    });
-  });
-
-  describe('removeAllPlugins', () => {
-    it('should remove all plugins', async () => {
-      // Arrange
-      vi.mocked(PluginModel.clear).mockResolvedValue(undefined);
-
-      // Act
-      const result = await pluginService.removeAllPlugins();
-
-      // Assert
-      expect(PluginModel.clear).toHaveBeenCalled();
-      expect(result).toBe(undefined);
-    });
-  });
-
-  describe('updatePluginSettings', () => {
-    it('should update plugin settings', async () => {
-      // Arrange
-      const id = 'plugin-id';
-      const settings = { color: 'blue' };
-      vi.mocked(PluginModel.update).mockResolvedValue(1);
-
-      // Act
-      const result = await pluginService.updatePluginSettings(id, settings);
-
-      // Assert
-      expect(PluginModel.update).toHaveBeenCalledWith(id, { settings });
-      expect(result).toEqual(1);
-    });
-  });
-
   it('can parse the OpenAI plugin', async () => {
-    const manifest = pluginService['convertOpenAIManifestToLobeManifest'](OpenAIPlugin as any);
+    const manifest = toolService['convertOpenAIManifestToLobeManifest'](OpenAIPlugin as any);
 
     expect(manifest).toMatchSnapshot();
   });
