+++ src/services/plugin/client.ts
@@ -1,0 +1,44 @@
+import { LobeChatPluginManifest } from '@lobehub/chat-plugin-sdk';
+
+import { PluginModel } from '@/database/client/models/plugin';
+import { LobeTool } from '@/types/tool';
+import { LobeToolCustomPlugin } from '@/types/tool/plugin';
+
+export interface InstallPluginParams {
+  identifier: string;
+  manifest: LobeChatPluginManifest;
+  type: 'plugin' | 'customPlugin';
+}
+
+export class PluginService {
+  installPlugin = async (plugin: InstallPluginParams) => {
+    return PluginModel.create(plugin);
+  };
+
+  getInstalledPlugins = () => {
+    return PluginModel.getList() as Promise<LobeTool[]>;
+  };
+
+  uninstallPlugin(identifier: string) {
+    return PluginModel.delete(identifier);
+  }
+
+  async createCustomPlugin(customPlugin: LobeToolCustomPlugin) {
+    return PluginModel.create({ ...customPlugin, type: 'customPlugin' });
+  }
+
+  async updatePlugin(id: string, value: LobeToolCustomPlugin) {
+    return PluginModel.update(id, value);
+  }
+  async updatePluginManifest(id: string, manifest: LobeChatPluginManifest) {
+    return PluginModel.update(id, { manifest });
+  }
+
+  async removeAllPlugins() {
+    return PluginModel.clear();
+  }
+
+  async updatePluginSettings(id: string, settings: any) {
+    return PluginModel.update(id, { settings });
+  }
+}
