+++ src/services/plugin/index.ts
@@ -1,0 +1,5 @@
+import { PluginService } from './client';
+
+export type { InstallPluginParams } from './client';
+
+export const pluginService = new PluginService();
